:recycle: (refactor) clean up the bot main service
